# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StenUp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get users",
                "description": "Get list of users",
                "operationId": "user-list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create an user",
                "description": "Create an user",
                "operationId": "create-user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "email_verified_at": {
                                        "description": "Email verified date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "role_ids": {
                                        "description": "Role IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Invalid input or email taken"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/users/{id}/ban": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Ban an user",
                "description": "Ban an user",
                "operationId": "ban-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation with no content in return"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/users/{id}/unban": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Unban an user",
                "description": "Unban an user",
                "operationId": "unban-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation with no content in return"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete an user",
                "description": "Delete an user",
                "operationId": "delete-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation with no content in return"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Edit an user",
                "description": "Edit an user",
                "operationId": "edit-user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email_verified_at": {
                                        "description": "Email verified date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "role_ids": {
                                        "description": "Role IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/users/collection:batchDelete": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Delete selected users",
                "description": "Delete selected users",
                "operationId": "delete-user-batch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "description": "Users' IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation with no content in return"
                    },
                    "422": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/users/registered_user_stats": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get registered user stats",
                "description": "Get registered user stats",
                "operationId": "registered-user-stats",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}